[FIX] 404 configuration not determined correctly
Due to the changes in the data structure, meaning that the domain is now
stored as a reference the 404 configuration was not correctly determined
based on the hostname. It did not work without any domain record set.

Now the configuration is correctly determined even if there is no domain
record present, means tx_naworkuri_helper::getCurrentDomain returns zero.