[FEATURE] Advanced page not found handling
The page not found handling is moved to seperate function so it can be
called from TYPO3 as a user function for page not found handling
configured in $GLOBALS['TYPO3_CONF_VARS']['FE']['pageNotFound_handling'].

Additionally it is distinguished wether the page is really not found or
the page is not accessible without being logged in. In case of a
non-accessible page a seperate configuration can be, but don't need to be
given. So it is possible to show a different errors for the not found and
not accessible cases.

An example configuration is added to the default configuration file.