[FEATURE] Improve page not found handling
The following applies to page not found and page not available handling:

Allow setting an integer id or string alias for the page not found page.
If the settings is an integer or the alias resolves to an integer do not
make a request to the configured value but set the value as the id
property in the TypoScriptFrontendController and let TYPO3 do the normal
rendering.

The status header is still set to the configured status value.