[BUGFIX] UpdateCacheHashWizard disabled to early
If the field migration (2.x to 3.x) haven been applied, the check
returned 0 records and the wizard was disabled.

If the count on the new `parameters` field is 0, it is checked, if the
old `params` field is still present and a count against this field is
made.

If both return 0, the wizard is disabled.